+++ src/indicators/jrc_agricultural_hotspots/utils/summary_agricultural_hotspots.R
@@ -5,8 +5,11 @@ box::use(
 
 box::use(
   src/utils/ai_summarizer,
-  src/utils/get_prompts
+  src/utils/get_prompts,
+  src/utils/get_manual_info
 )
+
+jrc_agricultural_hotspots <- "jrc_agricultural_hotspots"
 
 #' Add campaign info to cholera alerts
 #'
@@ -14,7 +17,17 @@ box::use(
 #'
 #' @export
 summary <- function(df_alerts, df_wrangled, df_raw) {
-  prompts <- get_prompts$get_prompts("jrc_agricultural_hotspots")
+  prompts <- get_prompts$get_prompts(jrc_agricultural_hotspots)
+
+  df_manual <- df_alerts |>
+    dplyr$filter(indicator_id == jrc_agricultural_hotspots) |>
+    dplyr$distinct(iso3, date) |>
+    dplyr$rowwise() |>
+    dplyr$mutate(
+      manual_info = get_manual_info$get_manual_info(iso3, jrc_agricultural_hotspots, date)
+    ) |>
+    dplyr$ungroup()
+
 
   df_summary <- df_alerts |>
     dplyr$select(
@@ -29,11 +42,16 @@ summary <- function(df_alerts, df_wrangled, df_raw) {
       comment_date <= date,
       date - comment_date <= 190
     ) |>
+    dplyr$left_join(
+      df_manual,
+      by = c("iso3", "date")
+    ) |>
     dplyr$group_by(
       iso3,
       location,
       date
     ) |>
+
     dplyr$summarize(
       info = paste(
         date_label,
@@ -44,6 +62,10 @@ summary <- function(df_alerts, df_wrangled, df_raw) {
       .groups = "drop"
     ) |>
     dplyr$mutate(
+      # COMBINA info con manual_info
+      info = paste(
+        info, manual_info, sep = "\n"
+      ),
       summary_long = purrr$map2_chr(
         .x = prompts$long,
         .y = info,